Crate fvm_ipld_encoding

Re-exports

pub use serde;

Modules

de

Generic data structure deserialization framework.

go_vec_visitor

ipld_block

repr

ser

Generic data structure serialization framework.

serde_bytes

A much simplified version of serde_bytes that:

strict_bytes

A much simplified version of serde_bytes that:

tuple

Structs

BytesDe

Wrapper for serializing and deserializing dynamic sized Bytes.

BytesSer

Wrapper for serializing slice of bytes.

Error

Error type for encoding and decoding data through any Forest supported protocol.

GoVecVisitor

Helper visitor to match Go’s default behaviour of serializing uninitialized slices as null. This will be able to deserialize null as empty Vectors of the type.

RawBytes

Raw serialized cbor bytes. This data is (de)serialized as a byte string.

Enums

CodecProtocol

CodecProtocol defines the protocol in which the data is encoded or decoded

Constants

CBOR

CBOR should be used to pass CBOR data when internal links don’t need to be traversable/reachable. When a CBOR block is loaded, said links will not be added to the reachable set.

DAG_CBOR

DagCBOR should be used for all IPLD-CBOR data where CIDs need to be traversable.

IPLD_RAW

RAW should be used for raw data.

Traits

CborDeprecated

Cbor utility functions for serializable objects

CborStore

Wrapper for database to handle inserting and retrieving ipld data with Cids

Functions

bytes_32

from_reader

Decode a value from CBOR from the given reader.

from_slice

Decode a value from CBOR from the given slice.

to_vec

Serializes a value to a vector.

to_writer

Encode a value as CBOR to the given writer.